Add configTASK_REENT_SUPPORT
This adds a void* to the pxTCB struct, intended to hold a reentrancy data structure.
The folling functions are called on task (1) allocation (2) deallocation (3) switch respectively:

uxPortTaskAllocateHook
vPortTaskDeallocateHook
vPortTaskSwitchHook

each of which takes a pointer to the void* in the pxTCB. The field may not be NULL when allocate is
called.